Implements the AppHeader component
