+++ React/.prettierrc
@@ -8,5 +8,5 @@
   "bracketSpacing": true,
   "jsxBracketSameLine": false,
   "arrowParens": "always",
-  "printWidth": 100
+  "printWidth": 120
 }
